Minor code cleanup
Renamed some methods
Name length cannot be longer than 15 characters
Spelling fix
Refactored getNetworkTime to send DateTime.Now() if no connection

+++ lib/util/formatter.dart
@@ -31,6 +31,6 @@ class Formatter {
   static int roundToInt(num value) {
     if (value is int) return value;
 
-    return value.round().toInt();
+    return value.round();
   }
 }
